requirements.txt
openai
google-generativeai
anthropic
python-dotenv